+++ src/sota_tools/garage_common.h
@@ -1,5 +1,6 @@
 #ifndef GARAGE_COMMON_H_
 #define GARAGE_COMMON_H_
+/** \file */
 
 /** Execution mode to run garage tools in. */
 enum class RunMode {
@@ -19,13 +20,13 @@ enum class RunMode {
  *  Copied here to avoid a dependency. We do not currently handle types 5-7, and
  *  UNKNOWN is our own invention for pseudo-backwards compatibility.
  *
- * @OSTREE_OBJECT_TYPE_FILE: Content; regular file, symbolic link
- * @OSTREE_OBJECT_TYPE_DIR_TREE: List of children (trees or files), and metadata
- * @OSTREE_OBJECT_TYPE_DIR_META: Directory metadata
- * @OSTREE_OBJECT_TYPE_COMMIT: Toplevel object, refers to tree and dirmeta for root
- * @OSTREE_OBJECT_TYPE_TOMBSTONE_COMMIT: Toplevel object, refers to a deleted commit
- * @OSTREE_OBJECT_TYPE_COMMIT_META: Detached metadata for a commit
- * @OSTREE_OBJECT_TYPE_PAYLOAD_LINK: Symlink to a .file given its checksum on the payload only.
+ * OSTREE_OBJECT_TYPE_FILE: Content; regular file, symbolic link
+ * OSTREE_OBJECT_TYPE_DIR_TREE: List of children (trees or files), and metadata
+ * OSTREE_OBJECT_TYPE_DIR_META: Directory metadata
+ * OSTREE_OBJECT_TYPE_COMMIT: Toplevel object, refers to tree and dirmeta for root
+ * OSTREE_OBJECT_TYPE_TOMBSTONE_COMMIT: Toplevel object, refers to a deleted commit
+ * OSTREE_OBJECT_TYPE_COMMIT_META: Detached metadata for a commit
+ * OSTREE_OBJECT_TYPE_PAYLOAD_LINK: Symlink to a .file given its checksum on the payload only.
  *
  * Enumeration for core object types; %OSTREE_OBJECT_TYPE_FILE is for
  * content, the other types are metadata.